Remove the code-commit related code
