+++ lib/constants.ts
@@ -22,7 +22,6 @@ export const PROVISION_TYPES = {
 export const AUTH_SERVICE_NAME = "auth-service";
 export const BUILD_SERVICE_NAME = "build-service";
 export const EMULATORS_SERVICE_NAME = "emulators-service";
-export const CODE_COMMIT_SERVICE_NAME = "code-commit-service";
 export const MISC_SERVICE_NAME = "misc-service";
 export const ACCOUNTS_SERVICE_NAME = "accounts-service";
 export const PROJECT_SERVICE_NAME = "project-service";
